+++ .gitignore
@@ -1,3 +1,6 @@
+# yes
+.idea/
+
 ## Ignore Visual Studio temporary files, build results, and
 ## files generated by popular Visual Studio add-ons.
 ##
