+++ README.md
@@ -3,16 +3,12 @@
 This mod unlocks the possibility for more than 10 player to join in an Among Us lobby.
 The official servers won't allow for it though, so you will have to host an [Impostor](https://github.com/Impostor/Impostor) custom server.
 
-
 ## Limitations
 By itself this mod doesn't allow to choose more than 10 players or more than 3 impostors in the game creation menu.
 You will have to change those values by other means (using memory editing or for example [All Of Us Impostor plugin](https://github.com/XtraCube/AllOfUsBot))
 
 ## Installation
 See: [client installation](https://github.com/CrowdedMods/CrowdedMod/wiki/Client-Installation)
-
-## TO-DO
-- switch to [semantic versioning](https://semver.org/)
 
 ## Credits
 - [andry08/100-player-mod](https://github.com/andry08/100-player-mod) - Original author of this plugin. RayanTheBest/100-player-mod was a mirror of this, now we have an official link.
