--- RemovePlayerLimit.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30524.135
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RemovePlayerLimit", "RemovePlayerLimit\RemovePlayerLimit.csproj", "{AB79A6CE-60D8-4F2F-9797-7F92A3EFCBF2}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AB79A6CE-60D8-4F2F-9797-7F92A3EFCBF2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AB79A6CE-60D8-4F2F-9797-7F92A3EFCBF2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AB79A6CE-60D8-4F2F-9797-7F92A3EFCBF2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AB79A6CE-60D8-4F2F-9797-7F92A3EFCBF2}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {767BDF72-7C0E-490F-8C81-927940B7CCFA}
-	EndGlobalSection
-EndGlobal
